Don't clean up old options unless this is a multisite install

+++ upload-scanner/class-upload-scanner-plugin.php
@@ -130,17 +130,19 @@ class Upload_Scanner_Plugin {
 		    update_site_option( 'upload-scanner_version', '1.2' );
 		    
 		    // Delete single site options
-		    delete_option( 'upload-scanner_use_clamav' );
-		    delete_option( 'upload-scanner_use_command' );
-		    delete_option( 'upload-scanner_command' );
-		    delete_option( 'upload-scanner_onfail_email_admin' );
-		    delete_option( 'upload-scanner_onfail_quarantine_file' );
-		    delete_option( 'upload-scanner_quarantine_folder' );
-		    delete_option( 'upload-scanner_onfail_send_406' );
-		    delete_option( 'upload-scanner_version' );
-		    delete_option( 'upload-scanner_onfail_email' );
-		    delete_option( 'upload-scanner_onfail_log_message' );
-		    delete_option( 'upload-scanner_onfail_log_file' );
+		    if ( is_multisite() ) {
+			delete_option( 'upload-scanner_use_clamav' );
+			delete_option( 'upload-scanner_use_command' );
+			delete_option( 'upload-scanner_command' );
+			delete_option( 'upload-scanner_onfail_email_admin' );
+			delete_option( 'upload-scanner_onfail_quarantine_file' );
+			delete_option( 'upload-scanner_quarantine_folder' );
+			delete_option( 'upload-scanner_onfail_send_406' );
+			delete_option( 'upload-scanner_version' );
+			delete_option( 'upload-scanner_onfail_email' );
+			delete_option( 'upload-scanner_onfail_log_message' );
+			delete_option( 'upload-scanner_onfail_log_file' );
+		    }
 		}
 	}
