+++ upload-scanner/uninstall.php
@@ -16,14 +16,16 @@ delete_site_option( 'upload-scanner_onfail_log_message' );
 delete_site_option( 'upload-scanner_onfail_log_file' );
 
 // Delete 1.1 and lower options that may be hanging around
-delete_option( 'upload-scanner_use_clamav' );
-delete_option( 'upload-scanner_use_command' );
-delete_option( 'upload-scanner_command' );
-delete_option( 'upload-scanner_onfail_email_admin' );
-delete_option( 'upload-scanner_onfail_quarantine_file' );
-delete_option( 'upload-scanner_quarantine_folder' );
-delete_option( 'upload-scanner_onfail_send_406' );
-delete_option( 'upload-scanner_version' );
-delete_option( 'upload-scanner_onfail_email' );
-delete_option( 'upload-scanner_onfail_log_message' );
-delete_option( 'upload-scanner_onfail_log_file' );
+if ( is_multisite() ) {
+    delete_option( 'upload-scanner_use_clamav' );
+    delete_option( 'upload-scanner_use_command' );
+    delete_option( 'upload-scanner_command' );
+    delete_option( 'upload-scanner_onfail_email_admin' );
+    delete_option( 'upload-scanner_onfail_quarantine_file' );
+    delete_option( 'upload-scanner_quarantine_folder' );
+    delete_option( 'upload-scanner_onfail_send_406' );
+    delete_option( 'upload-scanner_version' );
+    delete_option( 'upload-scanner_onfail_email' );
+    delete_option( 'upload-scanner_onfail_log_message' );
+    delete_option( 'upload-scanner_onfail_log_file' );
+}
